se anadio un nuevo menu compras hechas, le permite al usuario saber sus compras realizadas, es una consulta a la base de datos de los elementos con su id que esten ela tabla prodorden, hay que diseñar los elelmentos del carrito, tendran la misma estetica de compras hechas pero sin el add cantidad

+++ orders/models.py
@@ -75,7 +75,7 @@ class carritoCompras(models.Model):
     id_dueno = models.ForeignKey(User,on_delete=models.CASCADE,related_name="id_dueno_prod_carrito",null=False)
     subtipo_prod_car = models.CharField(max_length=200,null=False)
     tamano_prod_car = models.CharField(max_length=100,null=False)
-    precio_prod_car = models.IntegerField(null=False)
+    precio_prod_car = models.DecimalField(max_digits=10,decimal_places=2,null=False)
     toppings_prod_car = models.CharField(max_length=200,null=True)
     fecha_prod_car = models.DateTimeField(auto_now=False,auto_now_add=True)
     def __str__(self):
